third_party/WebKit/LayoutTests/platform/win/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/cors-preflight2-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&ACAMethods=PUT should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
